refactor code, add documentation

+++ ipc/ipc.cpp
@@ -13,44 +13,50 @@ static bool create_fifo() {
     return true;
 }
 
-static int open_fifo() {
-    int fifo = open(FIFO1, O_RDWR);
-    return fifo;
-}
-
-bool IPC::send_data(std::unique_ptr<char[]> buffer, uint64_t buffer_size) {
-    create_fifo();
+int IPC::create_ipc() {
+    if(!create_fifo())
+        return false;
 
     int fifo = open (FIFO1, O_RDWR);
     if(fifo == -1) {
         unlink(FIFO1);
-        return false;
+        return -1;
     }
     
+    return fifo;
+}
+
+
+bool IPC::send_data(int fifo, std::unique_ptr<char[]> buffer, uint64_t buffer_size) {
+    std::cout << "SENDING DATA BYTES: " << buffer_size << std::endl;
     if(write(fifo, &buffer_size, sizeof(uint64_t)) != sizeof(uint64_t) ||
-        write(fifo, buffer.get(), buffer_size) != buffer_size)
+        write(fifo, buffer.get(), buffer_size) != buffer_size) 
         return false;
 
+    std::cout << "DATA SEND COMPLETE" << std::endl;
     return true;
 }
 
-std::unique_ptr<char[]> IPC::receive_data(uint64_t &buffer_size) {
+std::unique_ptr<char[]> IPC::receive_data(int fifo, uint64_t &buffer_size) {
+    std::cout << "RECEIVING DATA" << std::endl;
     buffer_size = 0;
-    create_fifo();
-
-    int fifo = open(FIFO1, O_RDWR);
-    if(fifo == -1) {
-        unlink(FIFO1);
-        return nullptr;
-    }
-
     if(read(fifo, &buffer_size, sizeof(uint64_t)) != sizeof(uint64_t))
         return nullptr;
-    
+
+    std::cout << "RECEIVING DATA BYTES: " << buffer_size << std::endl;
     auto buffer = std::make_unique<char[]>(buffer_size);
 
     uint64_t data_read = 0;
-    while((data_read += read(fifo, buffer.get() + data_read, buffer_size)) != buffer_size);
+    while((data_read += read(fifo, buffer.get() + data_read, buffer_size)) != buffer_size) {
+        std::cout << "RECEIVED SOME DATA: " << data_read << std::endl;
+    }
+
+    std::cout << "RECEIVING DATA COMPLETE" << std::endl;
 
     return std::move(buffer);
+}
+
+void IPC::destroy_ipc(int fifo) {
+    close(fifo);
+    unlink(FIFO1);
 }
